Call monitor input selection script, upon return to main screen
This is implemented as a complete hack.

The current  enter_hook is set in the configuration to call the
script properly when a client screen is entered, but unfortunately
that isn't executed when the mouse returns to the main screen.

I don't have time to integrate the proper change so I just hacked
in this direct call to do what I need.